+++ common-db/change-log.md
@@ -14,6 +14,9 @@
 13. JdbcUtils加入存储过程的参数设置和存储过程的OUT Param的获取
 14. ignorePolicy重命名为ignoreStrategy
 15. 完善CatalogMetadata,SchemaMetadata,DatabaseMetadata
+16. 修复SelectColumnElement未设置distinct的问题
+17. SqlSortBuilder实现asc(String tableAlias, Field field),desc(String tableAlias, Field field)
+18. SqlSelectColumnsBuilder加入setColumnAliasPrefix(boolean)方法
 
 # 0.5.5 2022-8-11
 1. Dialect加入getKeywordLike(QueryPolicy)、getKeywordEq(QueryPolicy)、keywordsCase()方法